+++ src/pages/EditPage/index.js
@@ -456,8 +456,8 @@ const EditPage = ({ match }) => {
           district: districtSelected.value,
           city: "Cascavel",
           state: "PR",
-          latitude: parseInt(latitude),
-          longitude: parseInt(longitude),
+          latitude: parseFloat(latitude),
+          longitude: parseFloat(longitude),
         },
         imagens: uploadedFiles.map((file) => ({
           name: file.file ? file.file.name : file.name,
